sales-agent-labs/presgen-assess/requirements.txt
# PresGen-Assess Dependencies

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.36
alembic==1.12.1
# Skip asyncpg for testing - use sqlite for local testing
# asyncpg==0.29.0

# Vector database and embeddings
chromadb==0.4.15
openai==1.3.7

# Document processing
PyPDF2==3.0.1
pypdf==3.16.2  # Updated PDF processing library
python-docx==1.1.0
aiofiles==23.2.0  # Async file operations
nltk==3.8.1  # Natural language processing for text analysis

# Data validation and serialization
pydantic==2.11.9
pydantic-settings==2.11.0

# HTTP client for integrations
httpx==0.25.2

# Google API clients
google-api-python-client==2.183.0

# Configuration and environment
python-dotenv==1.0.0

# Text processing and chunking
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Async task processing
celery==5.3.4
redis==5.0.1

# Monitoring and observability
prometheus-client==0.19.0

# Testing framework
pytest==7.4.3
pytest-asyncio==0.20.3
pytest-mock==3.12.0

# Utilities
tenacity==8.2.3

# Authentication and security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
